+++ R/configurations.R
@@ -1,6 +1,19 @@
 # Configuration Environment, to be used as needed.
 .config <- new.env(parent = emptyenv())
 
+# REMOVE ONCE UPDATE GOES THROUGH
+.config$agreed_to_upcoming_loggit_updates <- FALSE
+#' Agree to upcoming loggit changes
+#'
+#' This function takes no arguments, but must be run before calling `loggit()`
+#' in a session. Due to the large, breaking changes coming in loggit v2.0, users
+#' are required to acknowledge that they have seen the error message before
+#' continuing to use this package's logging facilities.
+#'
+#' @export
+agree_to_upcoming_loggit_updates <- function() {
+  .config$agreed_to_upcoming_loggit_updates <- TRUE
+}
 
 #' Set Log File
 #'