+++ R/handlers.R
@@ -18,7 +18,9 @@ NULL
 #' 
 #' @inheritParams base::message
 #' 
-#' @examples if (2 < 1) message("Don't say such silly things!")
+#' @examples
+#' agree_to_upcoming_loggit_updates()
+#' if (2 < 1) message("Don't say such silly things!")
 #'
 #' @export
 message <- function(..., domain = NULL, appendLF = TRUE, .loggit = TRUE,
@@ -34,7 +36,9 @@ message <- function(..., domain = NULL, appendLF = TRUE, .loggit = TRUE,
 #'
 #' @inheritParams base::warning
 #' 
-#' @examples if (2 < 1) warning("You may want to review that math, and so this is your warning")
+#' @examples
+#' agree_to_upcoming_loggit_updates()
+#' if (2 < 1) warning("You may want to review that math, and so this is your warning")
 #' 
 #' @export
 warning <- function(..., call. = TRUE, immediate. = FALSE, noBreaks. = FALSE, 
@@ -52,7 +56,9 @@ warning <- function(..., call. = TRUE, immediate. = FALSE, noBreaks. = FALSE,
 #' 
 #' @inheritParams base::stop
 #' 
-#' @examples if (2 < 1) stop("This is a completely false condition, which throws an error")
+#' @examples
+#' agree_to_upcoming_loggit_updates()
+#' if (2 < 1) stop("This is a completely false condition, which throws an error")
 #' 
 #' @export
 stop <- function(..., call. = TRUE, domain = NULL, .loggit = TRUE,